+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This is the changelog for Tarmac.
 
 - Add `exists`, `isfile`, and `isdir` condition helpers
 - Add `changed` condition helper
+- Add `skipped` condition helper
 
 ## [0.1.7]